separate class for managing author
