+++ Classes/app.rb
@@ -1,12 +1,16 @@
 require 'date'
-require_relative 'author'
 require_relative 'game'
+require_relative 'author_app'
+require_relative 'utils/util'
 
 class App
   def initialize
     @all_things = []
+
+    # All subclass of item class (Book, MusicAlbum, Movie, and Game) should be push here
     @items = []
-    @authors = []
+
+    @author_manager = AuthorApp.new
   end
 
   def create_book
@@ -19,38 +23,25 @@ class App
   end
 
   def list_authors
-    if @authors.empty?
-      check_and_create_author
-    else
-      puts "\nList of Authors:"
-      @authors.each_with_index do |author, i|
-        puts "  #{i} | Name: #{author.first_name} #{author.last_name}. id:#{author.id}"
-      end
-      puts ''
-    end
+    @author_manager.list_authors
   end
 
   def list_items
     if @items.empty?
-      puts 'Items is empty'
+      puts "Item is empty\n\n"
     else
       puts "\nList of Items:"
       @items.each_with_index do |item, i|
-        puts "  #{i} | [#{item.class.name}] Publish Date: #{item.publish_date}. id: #{item.id}"
+        author_first_name = item.author.first_name && ''
+        puts "  #{i} | [#{item.class.name}] - Author: #{author_first_name} "
+        print "Publish Date: #{item.publish_date}. id: #{item.id}"
       end
       puts ''
     end
   end
 
   def create_author
-    first_name = get_non_empty_input('Author first name')
-    last_name = get_non_empty_input('Author last name')
-
-    author = Author.new(first_name, last_name)
-    @authors << author
-
-    puts "Author #{author.first_name}#{author.id} created successfully!"
-    author
+    @author_manager.create_author
   end
 
   def create_item
@@ -74,16 +65,17 @@ class App
   end
 
   def create_game
+    return unless @author_manager.check_and_create_author
+
     last_played_at = get_date_input('Last played date')
     multiplayer = get_boolean_input('Is it multiplayer?')
     publish_date = get_date_input('Publish date')
 
     loop do
-      check_and_create_author
-
       list_authors
+
       author_id = get_non_empty_input('Select Author by ID listed above').to_i
-      author = @authors.find { |i| i.id == author_id }
+      author = @author_manager.authors.find { |i| i.id == author_id }
 
       if author.nil?
         puts 'Invalid author ID. Please choose a correct author ID listed above or enter "exit" to go back to the Menu.'
@@ -98,64 +90,5 @@ class App
       end
     end
   end
-
-  private
-
-  def get_non_empty_input(prompt)
-    input = nil
-    loop do
-      print "#{prompt}: "
-      input = gets.chomp.strip
-      break unless input.empty?
-
-      puts "\"#{prompt}\" cannot be empty. Please try again."
-    end
-    input
-  end
-
-  def get_date_input(prompt)
-    loop do
-      print "#{prompt} (YYYY-MM-DD): "
-      input = gets.chomp.strip
-
-      case
-      when /^\d{4}-\d{2}-\d{2}$/.match?(input)
-        return input
-      when input.casecmp('now').zero? || input.casecmp('today').zero?
-        return Date.today
-      when input.empty? || input.casecmp('nil').zero? || input.casecmp('null').zero?
-        return nil
-      else
-        puts 'Invalid date format. Please enter a date in the format YYYY-MM-DD.'
-      end
-    end
-  end
-
-  def get_boolean_input(prompt)
-    loop do
-      print "#{prompt} (Y/N): "
-      input = gets.chomp.strip.downcase
-
-      case input
-      when 'y'
-        return true
-      when 'n'
-        return false
-      else
-        puts 'Invalid input. Please enter Y/N.'
-      end
-    end
-  end
-
-  def check_and_create_author
-    return unless @authors.empty?
-
-    puts 'No author is currently in the list'
-    choice = get_boolean_input('Do you want to create an author?')
-    return unless choice
-
-    create_author
-  end
-
   # you can add your required def here
 end
